[API-10] Refactor on Message Service
