+++ controllers/messages.py
@@ -1,9 +1,8 @@
 from models.authentication import Authenticator
 from models.request import ClientRequest
-from models.constants import StatusCode
-from models.message import Message
+from services.message import MessageService
 
-from flask import request, jsonify
+from flask import request
 from run import app
 
 import logging
@@ -15,6 +14,6 @@ def get_messages():
     logger.info("Attempting to get messages")
     req = ClientRequest(request)
     Authenticator.authenticate(req.authentication_data())
-    messages = Message.query.all()
+    messages = MessageService.query.all()
     contents = [message.text_content() for message in messages]
     return "Mensajes: " + ", ".join(contents)
